requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
google-cloud-firestore==2.14.0
google-auth-oauthlib==1.2.0
